Se arregla parseo en descarga de productos_prov

+++ lib/modelsEmiWeb/get_productos_prov_by_id_emi_web.dart
@@ -44,7 +44,7 @@ class Payload {
     final int idCatTipoProducto;
     final String marca;
     final String descripcion;
-    final int costoUnidadMedida;
+    final double costoUnidadMedida;
     final int tiempoEntrega;
     final int idUnidadMedida;
     final bool archivado;
